+++ CHANGELOG.md
@@ -9,10 +9,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Original Weight Weighting scheme
 - Add Is Valid validation in risk and return engines
+- Add Area Plot for weights
+- Allow user to set weight constraints on asset level
+- Scaled HRP
 ### Fixed
 - Fixed Monthly vs. Daily Bug in Covariance Matrix
 ### Changed
 - Only update return and risk engine once, not in every strategy
+- Move Stop-Loss functionality in seperate folder
 ### Removed
 
 ## [1.5.1] - 2023-12-18